client/sync/v5: Split rooms.rs into bump_stamp and heroes submodules.
